requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
cssselect==1.1.0
fake-useragent==0.1.11
Flask==2.0.2
Flask-Cors==3.0.10
Flask-MySQL==1.5.2
Flask-RESTful==0.3.9
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
lxml==4.6.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
parse==1.19.0
protobuf==3.18.0
pycparser==2.21
pyee==8.2.2
PyMySQL==1.0.2
pyppeteer==0.2.6
pyquery==1.4.3
python-decouple==3.5
pytz==2021.3
requests==2.26.0
requests-html==0.10.0
six==1.16.0
soupsieve==2.2.1
tqdm==4.62.3
urllib3==1.26.7
w3lib==1.22.0
waitress==2.0.0
websockets==9.1
Werkzeug==2.0.2
zipp==3.5.1
